graph, mock, store: Remove unnecessary Store unsubscribe logic
A more elegant way implemented in this commit is to drop Store
subscriptions when their receiving end has disappeared.

We're currently doing this kind of clean up only when trying to
write changes to subscription streams. This would allow clients
to open A LOT of subscriptions then close the connections and the
streams would stay alive as long as there are no new entity
changes to be sent. This is bad.

A good thing to do in addition would be to run a separate task
that, using a timeout, periodically calls `poll_ready()` on
subscription senders and removes all subscriptions for which the
sender returns `Err(SendError(_))` (see
https://docs.rs/futures/0.1/futures/sync/mpsc/struct.Sender.html#method.poll_ready
for more info).