json-rpc: First version of auth
If the `GRAPH_MASTER_TOKEN` is not set,
no auth will be checked. If it is set, then first
to authorize the deployment of a subgraph the
`subgraph_authorize` endpoint must be
called with a `name` parameter which will
return a fresh token that must be sent
along any deploy or remove requests for
a subgraph with that name.
Names that have never been authorized are
of course rejected.
All auth tokens must be provided
with the header `Authorization: Bearer TOKEN`.

Right now removal by id doesn't make sense
since we need the name to authorize,
but it's still supported by the rest of the codebase
since we don't yet know what will be the final
story for auth.