runtime: Fix `ipfs.cat`
The ipfs client must be run in the context
of a tokio runtime, so using `wait` was not
working.

In order to run a task
in the main runtime from the wasmi thread,
the WasmiModule now takes
a `task_sink` to which it sends futures
which need to be run in the runtime.

Adds a test for `ipfs.cat` and refactors
the module tests.