Remove 'subgraph_id' from StoreEvent
The field StoreEvent.subgraph_id suggests that all changes in theevent
relate to the same subgraph. That is generally not true, as the changes for
a 'real' subgraph always carry a change for the 'subgraphs' subgraph,
relating to changing its block pointer.

Since we don't really need the subgraph_id field on a StoreEvent, remove
the subgraph_id and accept that the changes of a StoreEvent may relate to
multiple subgraphs. That is also the only way to build filters on
StoreEventStreams that have to combine multiple StoreEvents into one (like
time-based throttling will have to)