store/postgres: pass large store_event notifications through a separate table
We create a large_notifications table. Notification payloads on the
store_events channel bigger than 8000 bytes are put into this table, and
the id of the entry in that table is sent instead.

The NotificationListener is responsible for transparently unpacking
payloads that go through this table.

Fixes https://github.com/graphprotocol/graph-node/issues/768