graph/components: add a Store.create_subgraph_deployment function
For now, it does exactly what we've always done when creating a subgraph
deployment. We need the registrar to tell the store that it's creating a
subgraph deployment explicitly, so the store can do more sophisticated
setup.

Note that there is no need for the parallel 'remove_subgraph_deployment'
since we never remove subgraph deployments.