graph: Add `TryFromValue` and `ValueMap` helper traits
`TryFromValue` can be used to try to convert a `graphql_parser::query::Value`
into a different type. It is like `TryFrom`, just with the `Error` type
fixed to be `failure::Error`.

`ValueMap` is a helper to pull out optional and required values from a
collection by name/key, and convert them using `TryFromValue` on the fly.

Both of these are convenient helpers to go from GraphQL results to
typed entities such as those in `graph::data::subgraph::schema`.