store/postgres: Refactor access to the entities table into its own module
We will need to be able to deal with entities stored in the current
monolithic entitites table, as well as with entities stored in per-subgraph
schemas. This is the first step in that direction, in that it lets us
abstract the differences between the two stores away.

Unfortunately, Diesel forces us to factor out entire queries; it's not
practical to just factor out access to each table since passing around
partial queries ('select _ from public.entities where subgraph=$1 ...') is
not practical because of the intermediate types that one has to deal
with. They are generally impossible to unify between the above partial
select and 'select _ from sgd1.entities' for split-out entities tables.