store/postgres: move responsibility for attr index naming to the store
The store is now responsible for what attribute indexes are called. In the
case of split entities tables, this also helps ensure that we do not run
into problems if the user's entity name and attribute name are so long that
we run over Postgres' 63 character limit on the length of the identifier.