store/postgres: Add `create_history_event` method
This method allows to create a history event to be created explicitly
before applying entity operations. Right now, this event is implicitly
created in the `subgraph_log_entity_event` trigger. For the subgraph of
subgraphs - and in the future, all subgraphs - we want to create the
entity history records directly from Rust though. For that we need to
be able to create the history event upfront.