store/postgres: Denormalize the entities count
We now store the number of entities in a subgraph as the `entityCount`
attribute in the SubgraphDeployment. If the SubgraphDeployment does not
have an `entityCount` yet, it is computed the first time it is requested,
which also provides a convenient way to force a recount by removing the
`entityCount` attribute.

Fixes https://github.com/graphprotocol/graph-node/issues/835