store/postgres: Do not demand that uncled blocks have a parent in the db
It is possible for us to pull an uncled block into the database that does
not have a valid parent on the chain. If such a block was within
ancestor_count of a new candidate head, we would refuse to advance the head
because of the missing parent, even though that parent is not needed.

This was ultimately caused by relying on block numbers to determine which
blocks must have a parent in the database. We now perform this check by
following the (hash, parent_hash) chain of blocks back from the candidate
head, rather than considering all blocks whose numbers are less than
ancestor_count from the candidate head.

Fixes https://github.com/graphprotocol/graph-node/issues/983