store/postgres: Reduce locking when cleaning up large notifications
The locks we need to acquire to clean up old large notifications can make a
transaction wait for one (or several) slow transactions, stalling the whole
system, or even lead to a deadlock.

With this change, we only clean up old large notifications that are not
currently locked by other transactions so that we will never wait for
them.

Fixes https://github.com/graphprotocol/graph-node/issues/987