graph: Use async client for sending Elasticsearch logs
Using a synchronous client can lead to `attempted to run an executor
while another executor is already running` errors like the one reported
on https://github.com/seanmonstar/reqwest/issues/541.

As part of switching to the asynchronous reqwest client, this commit
also reduces the time a lock on the queued logs is held. Instead of
locking until all logs have been sent, the queued logs are cloned and
the lock is released immediately.