*: Use one runtime host builder and block stream builder instead of set
Ethereum adapters are stored in a HashMap instead of Vector.
Use a single block stream and runtime host builder instead of a set.
They each have access to maps of stores and ethereum adapters,
one for each network.
Pass network_name to build() to select the appropriate store and
ethereum adapter for each subgraph.