main: Reduce noisiness of tokio contention check
Reduce frequency of check to every second and
downgrade log level to debug.

Also increase the wait time to 10ms.
In a production environment 1ms wait time
does signal an issue, but at that point we're
measuring contention in the OS scheduler,
and there are more suitable tools to measure
that than this log.