store: Do not get stuck on genesis block's missing parent
In Ganache setups, we pull in the genesis block as the only block in the
beginning. That block does not have a valid parent, and we must not include
it in our parent checking.

This chainge also adds tests for various issues we've had with
attempt_chain_head_update to avoid repeating the same bugs.

Fixes https://github.com/graphprotocol/graph-node/issues/1088