core, graphql, server, store: Add earliest and head block to deployments
Moving forward, we're going to want to know from which a block a
subgraph was indexed, or rather, which block is the earliest for
which we have subgraph data, as well as what the head block of the
chain the deployment is indexed from is.

This commit adds this information to the `SubgraphDeploymentEntity`.
Both the blocks are optional, as existing deployments won't have
them yet.