store: Deduplicate entity operations
For time-travel support, we need to avoid changing the same entitiy twice
for one block, and updating an entity to the state it is already in because
that leads to recording unnecessary history, increasing the size of entity
tables beyond what is needed, in some cases quite significantly.

These changes combine a list of entity operations so that we avoid the
above problems, but combining is only possible for lists of entity
operations that do not use the `Update` or `AbortUnless` variants of
`EntityOperation`; those are only used for subgraph metadata.