store: Check that identifiers are valid SQL names
We check that the names of the database schema, tables, and columns match
the regular expression /[A-Za-z][A-Za-z0-9_]*/ so we do not have to worry
about enclosing them in double quotes when interpolating them into SQL
queries. This should not pose a problem for users in practice, as that
coincides with the rules for GraphQL names.