store: Quote table and column names when creating relational schema
We make sure that the type and attribute names from GraphQL schemas are
safe to use in SQL, but users may use names that conflict with reserved
words in SQL; for example, a type 'User' will result in the creation of a
table 'user'. That causes a syntax error in Postgres, and we therefore need
to enclose these names in double-quotes during schema creation. Proper
queries do not have this problem since we have Diesel escape them as
identifiers already.