store: Make it clear that entities::Connection is for one subgraph
We only ever used one entities::Connection for one subgraph, plus the
metadata subgraph, though the code obscured that and made it look like the
same instance would be used for multiple subgraphs.

With these changes, we make it explicit that the Connection is for one
subgraph, and remove a lot of places where `subgraph_id` was passed
redundantly.