store: Optimize query listing entities
When listing entities we run the query

  select data, entity
    from entities
   where entity = ANY ('{Domain}')
   order by id asc
   limit 5

This query is very slow, since Postgres does not use the primary key index
on (entity, id) because of the 'any'. If there is only one entity type,
which is the case in most queries, we change the 'any' to a simple `entity
= 'Domain'` so that Postgres uses the primary key index. In one example,
that brings query time from 130s to 6ms.