core, graph: Use 60s timeout for loading subgraph files from IPFS
Before, whatever you set as the IPFS timeout was used not just for
`ipfs.cat` in mappings, but also for loading subgraphs. Now, when you
set the timeout to a short duration (like 2s), subgraphs would often
fail to start up. Restricting the time spent in `ipfs.cat` is reasonable
thing to do though, so the timeout used for `ipfs.cat` and subgraph
loading should be independent.

This change adds a `with_timeout` method to `LinkResolver` to create a
cloned resolver that has its own timeout. This is then used in the
`SubgraphRegistrar` to hard-code a 60s timeout for manifest resolution.
I don't feel this needs to be configurable at the moment.