core,datasource,graph,mock:Simplify start_block processing on startup
Do not check that the start_block contains the contracts creation
transaction as the subgraph owner is free to choose any block on chain.
Improve naming of function that processes the start_block:
validate_start_block() --> block_pointer_from_number().
Only process the earliest start_block from the manifest when creating a
SubgraphDeploymentEntity.