core, graph: Add `with_retries` modifier to `LinkResolver`
This allows to configure the link resolver so that it retries requests
indefinitely using our `retry` utility.

By default, this will be disabled but we'll want to enable it for the
resolver used when loading subgraphs manifest files and dynamic data
sources. The reason this would be good is that it would allow us to
tolerate intermittent IPFS issues during startup.