graphql: When prefetching, do not fall back to old query resolution
We want to make sure that we do not inadvertently resolve objects
individually when we already prefetched results so that we do not
inadvertently run into the reduced performance of the old
approach. Instead, we want queries to fail if prefetching does not return
all the results needed for resolution so that we can fix those bugs.