chain/ethereum: Add test for common ancestor identification bug
The bug is that the common ancestor detection assumes that any block
in the store is part of the local chain. However, since regular blocks
and ommers are both stored as `Block` entities, blocks looked up from
the store could also be ommers, and not part of the chain.

This test verifies that we handle this correctly. One of the next
commits will include a fix for the issue.