chain, graph, runtime: Switch to latest ethabi and rust-web3
This points to the graphprotocol forks of ethabi and rust-web3,
rebased against the latest upstream ethabi and rust-web3 master
branches.

This requires a few other dependencies to be updated and block
numbers to be changed from U128 to U64. To subgraphs this change
should not matter, as to them the number is a `BigInt` anyway.