graphql: Simplify the join logic and fix a bug
The logic for actually joining parents and children was overly complicated,
and for situations where the children were an interface caused repeating
the children for each parent as many times as the interface had
implementations.

Since we have the parents in memory already, and since parent_id's are
unique across all the parents that can possibly appear in a list of
parents, we can distribute children into their parents simply by looking at
the parent_id that the EntityQuery reports for each child already.

Fixes https://github.com/graphprotocol/graph-node/issues/1438