store: Check block hash against subgraph's network in block_number
Before, we were checking against the store's network; that is not correct
as queries use a store that has nothing to do with the subgraph's
network. That would lead to siutations where looking up a Ropsten hash for
a Ropsten subgraph would try to find the hash in mainnet and not find it
even though it was a valid Ropsten hash.