store: Produce correct empty array in FilterWindow.expand_parents
If a window connected children to parents through a list of child id's
stored in the parent, and if all parents in a query had an empty list, we
would generate `array[]::text[][]` which is wrong since we need to produce
a two-dimenional array. It does not seem possible to produce an entirely
empty two-dimensional array (a 0x0 matrix), instead we produce a 1x1 matrix
with an entry `null` as `array[array[null]]::text[][]`.