store: Make sure metadata gets reverted only for one subgraph
The way we used to revert metadata when a block was reverted was not
correct: the reversion would also revert metadata for unrelated subgraphs,
leading to potential loss of dynamic data sources.

This code now makes sure we only revert metadata for the subgraph for which
we are reverting a block.