store: Work around tables having a text column where they should have enums
We had a bug that if the GraphQL schema for a subgraph used an enum before
it was defined, the generated database schema would have a 'text' column
instead of an 'enum' column. Since that bug was (inadvertently) fixed in
commit 3d23bccf, queries against such tables would fail if they tried to
filter by the presumptive enum attribute.

With this commit, we check whether we have a text column where we should
have an enum, and then switch the internal table representation to treat
the enum as 'String' and not as an enum when querying or storing data.