graph: Correctly calculate the relative effort of a query
We used to compare the average execution time of a query to the average
total execution time, but that comparison is meaningless, and does not
indicate how much of the total workload is caused by that query - it's
entirely possible to have a total average execution time of 50ms and a
single query that takes 5s on average.

Rather than using averages, we now use the total time that a query took
during the window and compare it to the total time for all queries to
determine the fraction of the total effort caused by that query.