store: Allow setting which tables should use redundant block_range clauses
For tables that have a low ratio of distinct entities to overall rows,
i.e., where entities are updated frequently, using redundant block_range
clauses that make the BRIN index on block_range usable seems to be
beneficial.

Ultimately, we would like to automatically detect such tables, but to make
sure that this is in fact a useful heuristic, we allow setting which tables
to treat special in the environment to make it possible to test this
hypothesis under load.