store: Do not crash the node on invalid StoreEvent
It is possible to receive invalid store events when a node is starting up
because Postgres still has events sent by a previous version of the
code. In that case, the best we can do is ignore these events.

Once we've seen enough valid events, go back to panicing on invalid events.